speech-api/SpeechSynthesis-speak-without-activation-fails.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<p>Some body text.
<script>
var t = async_test('speechSynthesis.speak requires user activation');
const utter = new SpeechSynthesisUtterance('1');
utter.onerror = t.step_func_done((e) => {
  assert_equals(e.name, "not-allowed");
});
utter.onend = t.step_func_done(() => assert_unreached());
speechSynthesis.speak(utter);
</script>
